Use 'object=groups' instead of 'scope=groups'.
The latter has a particular meaning that is incorrect for the current
context. Specifically, 'scope=groups' means that activity will be
returned that are linked to the groups of which the displayed user
is a member. Instead, we want to limit to items with the 'groups'
component, which is what 'object=groups' means.